Vari cambiamenti grafici e aggiunta modifica account ancora non funzionante

+++ src/main/java/com/project/just_meet/web/EventController.java
@@ -5,8 +5,13 @@ import com.project.just_meet.model.User;
 import com.project.just_meet.service.event.EventService;
 import com.project.just_meet.service.user.UserService;
 import com.project.just_meet.validator.EventValidator;
+
 import java.security.Principal;
+import java.util.HashSet;
+import java.util.Set;
+
 import javax.transaction.Transactional;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -40,6 +45,7 @@ public class EventController {
 			return "eventCreation";
 
 		User u = userService.findByUsername(user.getName());
+		
 		eventForm.getUsers().add(u);
 		u.getEvents().add(eventForm);
 		eventService.save(eventForm);
@@ -62,16 +68,40 @@ public class EventController {
 	public String myEvents(Model model, Principal user) {
 		model.addAttribute("list", eventService.findAllByUsername(user.getName()));
 		model.addAttribute("id", new Event());
-
+		
 		return "/myEvents";
 	}
 
 	@Transactional
 	@PostMapping("/myEvents")
-	public String deleteEvent(@ModelAttribute("id") long id) {
+	public String deleteEvent(@ModelAttribute("id") long id, BindingResult bindingResult) {
+		if (bindingResult.hasErrors())
+			return "/myEvents";
+
+		Event event = eventService.findById(id);
+
+		for (User u : event.getUsers()) {
+			u.getEvents().remove(event);
+			userService.update(u);
+		}
+
 		eventService.deleteById(id);
 
 		return "/myEvents";
+	}
+
+	@GetMapping("/updateEvent")
+	public String updateEvent(Model model, @RequestParam long id) {
+		model.addAttribute("event", eventService.findById(id));
+
+		return "/updateEvent";
+	}
+
+	@PostMapping("/updateEvent")
+	public String updateEvent(@ModelAttribute("event") Event event) {
+		eventService.save(event);
+
+		return "redirect:/event?id=" + event.getId();
 	}
 
 	@GetMapping("/event")
@@ -83,21 +113,32 @@ public class EventController {
 	}
 
 	@PostMapping("/event")
-	public String addParticipation(@ModelAttribute("id") long id, Principal user) {
+	public String addParticipation(@ModelAttribute("id") long id, Principal user, BindingResult bindingResult) {
 		User u = userService.findByUsername(user.getName());
 		Event e = eventService.findById(id);
+		
 		e.getUsers().add(u);
 		u.getEvents().add(e);
+		
 		eventService.save(e);
-		userService.save(u);
+		userService.update(u);
 
-		return "redirect:/events";
+		return "redirect:/participations";
 	}
 
 	@GetMapping("participations")
 	public String partecipations(Model model, Principal user) {
-		model.addAttribute("participations", userService.findByUsername(user.getName()).getEvents());
+		Set<Event> events = userService.findByUsername(user.getName()).getEvents();
+		Set<Event> myEvents = new HashSet<Event>();
 
+		for (Event e : events)
+			if (e.getUsername().equals(user.getName()))
+				myEvents.add(e);
+
+		events.removeAll(myEvents);
+		
+		model.addAttribute("participations", events);
+		
 		return "/participations";
 	}
 }
